test: extract JSON-RPC integration helpers
